pyproject.toml
[project]
name = "forma"
version = "0.1.0"
description = "Python project bootstrapped by pyenv + uv"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "Jayleonc", email = "jayleonc@gmail.com" }]
dependencies = [
  "typer",
  "paddleocr",
  "paddlepaddle",
  "pillow",
  "tqdm",
  "rich",
  "rapidocr-onnxruntime",
  # OpenCV wheels used by paddlex/paddleocr currently require NumPy 1.x
  "numpy<2",
  "openai",
  "langchain",
  "langchain-openai",

  "pymupdf==1.26.3",
  "pymupdf4llm==0.0.27",

  "python-docx",
  "mammoth",
  "markdownify",
  "python-pptx",
  "pyyaml",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ipython>=8.0.0",
  "ruff", 
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]


[project.scripts]
forma = "forma.cli:app"
